LayoutTests/media/modern-media-controls/placard-support/placard-support-error-recover.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<video src="404.mp4" style="width: 320px; height: 240px;"></video>
<div id="shadow"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>PlacardSupport</code> behavior when we recover from an error.");

const shadowRoot = document.querySelector("div#shadow").attachShadow({ mode: "open" });
const media = document.querySelector("video");
const mediaController = createControls(shadowRoot, media, null);

media.addEventListener("error", () => {
    shouldBeTrue("mediaController.controls.showsPlacard");
    shouldBeTrue("mediaController.controls.children.includes(mediaController.controls.invalidPlacard)");

    media.addEventListener("loadstart", () => {
        shouldBeFalse("mediaController.controls.showsPlacard");
        debug("");
        shadowRoot.host.remove();
        media.remove();
        finishJSTest();
    });

    media.src = "../../content/test.mp4";
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
